+++ src/main/java/vendingmachine/utils/ProductFactory.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 import vendingmachine.model.Product;
+import vendingmachine.utils.exception.ProductException;
 import vendingmachine.view.InputView;
 import vendingmachine.view.OutputView;
 
@@ -18,7 +19,7 @@ public class ProductFactory {
 			List<Product> products = productsInfo.stream()
 				.map(Product::new)
 				.collect(Collectors.toList());
-			ExceptionUtils.validateDuplicatedName(products);
+			ProductException.validateDuplicatedName(products);
 			return products;
 		} catch (IllegalArgumentException IAE) {
 			OutputView.printError(IAE);
